backend/pyproject.toml
[project]
name = "yt-downloader-backend"
version = "0.1.0"
description = "Private YouTube downloader backend (FastAPI + yt-dlp + Valkey/RQ)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.9",
  "redis>=5.0.7",
  "rq>=1.16.2",
  "yt-dlp>=2024.12.13",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
